E-Commerce por IA/requirements.txt
fastapi==0.111.0
openai==1.34.0
postgrest==0.16.8
pydantic==2.7.4
PyYAML==6.0.1
PyYAML==6.0.1
supabase==2.5.1
twilio==9.1.1
